requirements.txt
tensorflow==2.5.0
google-cloud-storage==1.35.0
joblib==1.0.1
pandas==1.2.4
beautifulsoup4==4.9.3
lxml==4.6.3
scikit-learn==0.24.1
numpy==1.20.2